The internal pixel format of BiteSprite is now always 32bit RGBA.
The BMP loader has been made to assume the Bitesprite internal format. Color masks can be compared and
the pixel value can be converted between them. Although, for now they can only take masks that cover
4 bits at locations 0, 8, 16 and 24. A more advanced approach will not be implemented until needed.

The design and implementation of the code will be made alot simpler due to deciding the internal
format to be set at 32bit RGBA. If more modes is actually needed in the future, this can be changed
at that point in time.