Implemented colorkey for the spritesheet
Two functions, ColorKey and ColorKeyNorm allows you to pass the color you wish to use as color key for a sprite sheet. The key is given a sensitivity range, so any colors passing within it will be discarded in the fragment shader.

ColorKey take color values in the 8bit uint range, and ColorKeyNorm takes color values normalised to the range 0.0-1.0