Cleanup and stability fix.
Created the namespace Framebuffer and renamed the framebuffer functions to fit in it.

The framebuffer wasn't reloaded on resolution changes after Bite::Init(). Framebuffer::Reload() was created to solve this, and is called from the SetResolution function. It will only reload if BiteSprite has been initialized, and thus initialized glew. Not initializing glew ahead of this will cause access violations using any gl function which address is imported by glew.