Bugfix and tidy up
When a sprite was dropped it was given the ID=0, which resulted in the actual sprite with ID=0 being altered. Sprites now have an "alive" state, that when false won't allow any information to be passed to the batch. The ID is no longer erased when the sprite is dropped; the state "alive" is changed to false instead.

The CHECK_GL_ERROR has been improved with line and file information.