First steps to a palette implementation.
TGA loader generates a palette that tags along with the image data. When the sprite sheet is created it'll become the standard palette for that sheet. Currently no more palettes can be added, but that should be the next step. Sprite fragment shader has been made to utilize the new palette function, but only for a single palette.