Frames has position anchors now, and sprites can be flipped.
Instead of the set "top left" and "center" values when positioning a sprite, the frames store an anchor value. This is in preparation for animations.

Simple X and Y flipping can be done for sprites, so that the same sprite can face two directions, or in case the coordinates are in the an undesired direction you can flip your sprites.